signup, login and forgot password page seperated
