+++ components/auth/AuthMain.tsx
@@ -1,9 +1,11 @@
 import { useState } from 'react'
+
 import { Tabs } from '@space-metaverse-ag/space-ui'
 import { rgba } from '@space-metaverse-ag/space-ui/helpers'
 import styled from 'styled-components'
 import LoginForm from './LoginForm'
 import SignupForm from './SignupForm'
+import { useRouter } from 'next/router'
 
 const ModalWrapper = styled.div`
   width: 23.625rem;
@@ -39,22 +41,26 @@ enum AuthTabs {
   Signup = 'Signup',
 }
 
-const AuthMain = () => {
-  const [activeTab, setActiveTab] = useState(AuthTabs.Login)
+interface Props {
+  selectedTab: string,
+}
+
+const AuthMain: React.FC<Props> = ({ selectedTab }) => {
+  const router = useRouter() 
 
   const handleTabChange = (tab: string) => {
-    setActiveTab((tab as AuthTabs))
+    router.push(tab.toLowerCase())
   }
 
   return (
     <ModalWrapper>
       <TopSection>
         <WelcomeHeader>Welcome to Space</WelcomeHeader>
-        <Tabs tabs={['Login', 'Signup']} onChange={tab => handleTabChange((tab as string))} activeTab={activeTab} />
+        <Tabs tabs={['Login', 'Signup']} onChange={tab => handleTabChange((tab as string))} activeTab={selectedTab} />
       </TopSection>
       <FormSection>
-        {activeTab === AuthTabs.Login && <LoginForm />}
-        {activeTab === AuthTabs.Signup && <SignupForm finishSignup={() => setActiveTab(AuthTabs.Login)} />}
+        {selectedTab === AuthTabs.Login && <LoginForm />}
+        {selectedTab === AuthTabs.Signup && <SignupForm finishSignup={() => handleTabChange(AuthTabs.Login)} />}
       </FormSection>
     </ModalWrapper>
   )
