+++ components/auth/LoginForm.tsx
@@ -6,26 +6,20 @@ import {
   Checkbox,
   CheckboxStyles,
   TextInput,
-  Modal,
-  type ModalProps,
 } from "@space-metaverse-ag/space-ui";
 import {
   AuthError,
-  usePostForgotPasswordMutation,
   usePostLoginMutation,
 } from "api/auth";
 import styled from "styled-components";
 
-interface GeneralMessage {
-  data: { message: string }
-}
-
+import { useRouter } from "next/router";
 const Form = styled.form`
   gap: 1rem;
   display: flex;
   flex-direction: column;
 `;
-
+ 
 const FormButton = styled(Button)`
   width: 100%;
   margin: 0 auto;
@@ -35,11 +29,6 @@ const FormButton = styled(Button)`
 const Message = styled(Alert)`
   width: 100%;
   justify-content: center;
-`;
-
-const MessageWithMargin = styled(Message)`
-  margin-top: 12px;
-  margin-bottom: 12px;
 `;
 
 const ForgotLabel = styled(CheckboxStyles.Label)`
@@ -56,13 +45,9 @@ const LoginForm: React.FC = () => {
   const [password, setPassword] = useState<string>("");
   const [rememberMe, setRememberMe] = useState<boolean>(false);
 
-  const [emailForgot, setEmailForgot] = useState<string>("");
-  const [messageForgot, setMessageForgot] = useState<string>("");
-
   const formRef = useRef<HTMLFormElement>(null);
-  const modalRef = useRef<ModalProps>(null);
-  const formModalRef = useRef<HTMLFormElement>(null);
 
+  const router = useRouter()
   const [
     postLogin,
     {
@@ -73,15 +58,6 @@ const LoginForm: React.FC = () => {
       error: postLoginError,
     },
   ] = usePostLoginMutation();
-
-  const [
-    postForgotPassword,
-    {
-      isLoading: isPostForgotPasswordLoading,
-      isSuccess: isPostForgotPasswordSuccess,
-      isError: isPostForgotPasswordError,
-    },
-  ] = usePostForgotPasswordMutation();
 
   const handleLogin = useCallback(async () => {
     await postLogin({
@@ -197,7 +173,7 @@ const LoginForm: React.FC = () => {
             onChange={handleRememberMe}
             isChecked={rememberMe}
           />
-          <ForgotLabel onClick={() => modalRef.current?.opened()}>
+          <ForgotLabel onClick={() => router.push("/?forgotPasswordModal=true","/forgotPassword" ,{shallow: true})}>
             Forgot password?
           </ForgotLabel>
         </RememberAndForgot>
@@ -210,56 +186,6 @@ const LoginForm: React.FC = () => {
         />
       </Form>
 
-      <Form
-        ref={formModalRef}
-        onSubmit={async (e) => {
-          e.preventDefault();
-
-          await postForgotPassword({
-            email: emailForgot,
-          }).then((res) => {
-            setMessageForgot(
-              (res as GeneralMessage)?.data?.message ||
-                (res as { error: GeneralMessage })?.error?.data?.message
-            );
-          });
-        }}
-      >
-        <Modal
-          ref={modalRef}
-          actions={{
-            primary: {
-              color: "blue",
-              label: "Send Reset Password Link",
-              size: "large",
-              type: "submit",
-              disabled: isPostForgotPasswordLoading,
-            },
-          }}
-          close
-          outsideClick
-          shadow
-          size="small"
-          title="Recovery Password"
-        >
-          <TextInput
-            label="Email"
-            required
-            placeholder="email@mail.com"
-            type="email"
-            disabled={isPostForgotPasswordLoading}
-            value={emailForgot}
-            onChange={(e) => setEmailForgot(e.target.value)}
-          />
-
-          {(isPostForgotPasswordSuccess || isPostForgotPasswordError) && (
-            <MessageWithMargin
-              type={isPostForgotPasswordSuccess ? "success" : "error"}
-              text={messageForgot}
-            />
-          )}
-        </Modal>
-      </Form>
     </>
   );
 };
