+++ package.json
@@ -21,7 +21,7 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.0",
-    "@space-metaverse-ag/space-ui": "^0.3.0",
+    "@space-metaverse-ag/space-ui": "^0.5.0",
     "next": "13.0.2",
     "react": "18.2.0",
     "react-dom": "18.2.0",
