+++ pages/login/index.tsx
@@ -1,0 +1,7 @@
+import AuthWrapper from "components/auth/AuthWrapper";
+ 
+const Login: React.FC = () => {
+    return  <AuthWrapper selectedTab="Login"/> 
+}
+
+export default Login
