+++ pages/signup/index.tsx
@@ -1,0 +1,5 @@
+import AuthWrapper from "components/auth/AuthWrapper";
+ 
+const Signup: React.FC = () =>  <AuthWrapper selectedTab="Signup"/> 
+
+export default Signup
